requirements.txt
Flask
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
youtube-transcript-api
google-generativeai
fpdf2